FSAL_MDCACHE: fix duplicated put on creation of mdcache export
I hit an assert() while adding a vfs export after commit:
    FSAL: fix segfault while adding export for a non-exist subdirectory
Then I find that there is a duplicated fsal_put() on the failure code
path of mdcache_fsal_create_export(), because the fsal parameter is
expected to be freed in the caller where it gets alloced.

Also I add some frees for the mdcache_fsal_export which is alloced
right in mdcache_fsal_create_export().
Some out-of-date comment is corrected as well.

Change-Id: Iffee669ebbdb890ea1b35a494e365e4c40bc4718
Signed-off-by: Gui Hecheng <guihecheng@cmss.chinamobile.com>