Gluster: Avoid unloading the FSAL twice
While unloading FSAL_GLUSTER, if there is any reference
to active export, we call LogFATAL which again tries to unload
gluster resulting in double free.

This could happen when SIGKILL gets issued to the ganesha process
while an export is still being initialized and not yet attached to
the export list.

Change-Id: Ia69a42fd77e479501db65bf696289740989e0049
Signed-off-by: Soumya Koduri <skoduri@redhat.com>